Implement i2c state machine
This implements an i2c slave state machine using the USI peripheral.
Writes and reads can be done in bursts, and the register pointer will
wrap allowing continuous read/write.

For writes, the first byte after the slave address sets the register
pointer. The following bytes are the data to be written.
For reads, the current value of the register pointer is used to
start transferring data to the master.
The register pointer is auto-incremented after every data byte (read or
written)

TODO: Stop condition detection is done in the main loop. On the Attinyx5
and possibly others, we should be able to use a PCINT for this